Overwrite old dataset file, update tests and constructor
